Updated template to use int for part 2

+++ src/year2024/AoC202400.java
@@ -53,7 +53,7 @@ public class AoC202400 {
      * @return the result as an integer.
      * @throws FileNotFoundException if the file is not found
      */
-    public static long partTwo(String path) throws FileNotFoundException {
+    public static int partTwo(String path) throws FileNotFoundException {
         return 0;
     }
 }
